Moved buttonicon
