+++ packages/design-system-react-native/src/components/ButtonIcon/ButtonIcon.constants.ts
@@ -1,4 +1,4 @@
-import { IconSize } from '../../Icon';
+import { IconSize } from '../Icon';
 import type { ButtonIconProps } from './ButtonIcon.types';
 import { ButtonIconVariant } from './ButtonIcon.types';
