+++ packages/design-system-react-native/src/components/ButtonIcon/ButtonIcon.tsx
@@ -4,10 +4,10 @@ import React, { useMemo, useState } from 'react';
 import type { GestureResponderEvent } from 'react-native';
 import { Pressable, View } from 'react-native';
 
-import type { SpinnerProps } from '../../../temp-components/Spinner';
-import Spinner from '../../../temp-components/Spinner';
-import type { IconProps, IconColor } from '../../Icon';
-import Icon from '../../Icon';
+import type { SpinnerProps } from '../../temp-components/Spinner';
+import Spinner from '../../temp-components/Spinner';
+import type { IconProps, IconColor } from '../Icon';
+import Icon from '../Icon';
 import { DEFAULT_BUTTONICON_PROPS } from './ButtonIcon.constants';
 import type { ButtonIconProps } from './ButtonIcon.types';
 import {
